+++ trunk/game/physics/PlayerProjectileManager.cs
@@ -32,8 +32,8 @@ namespace AbrahmanAdventure.physics
                 {
                     playerSprite.ThrowBallCycle.Fire();
                     SoundManager.PlayFireBallSound();
-                    float xPosition = (playerSprite.IsTryingToWalkRight) ? playerSprite.RightBound + 0.5f: playerSprite.LeftBound - 0.5f;
-                    FireBallSprite fireBallSprite = new FireBallSprite(xPosition, playerSprite.TopBound + 0.33f, random);
+                    double xPosition = (playerSprite.IsTryingToWalkRight) ? playerSprite.RightBound + 0.5: playerSprite.LeftBound - 0.5;
+                    FireBallSprite fireBallSprite = new FireBallSprite(xPosition, playerSprite.TopBound + 0.33, random);
                     fireBallSprite.IsNoAiDefaultDirectionWalkingRight = playerSprite.IsTryingToWalkRight;
                     fireBallSprite.CurrentWalkingSpeed = playerSprite.CurrentWalkingSpeed + fireBallSprite.MaxWalkingSpeed;
                     fireBallSprite.CurrentJumpAcceleration = -30;
