+++ trunk/game/sprites/monsters/SnakeSprite.cs
@@ -29,7 +29,7 @@ namespace AbrahmanAdventure.sprites
         /// <param name="xPosition">x position</param>
         /// <param name="yPosition">y position</param>
         /// <param name="random">random number generator</param>
-        public SnakeSprite(float xPosition, float yPosition, Random random)
+        public SnakeSprite(double xPosition, double yPosition, Random random)
             : base(xPosition, yPosition, random)
         {
             GetLeft1Surface();
@@ -41,59 +41,59 @@ namespace AbrahmanAdventure.sprites
         #endregion
 
         #region Override Methods
-        protected override float BuildJumpingTime()
+        protected override double BuildJumpingTime()
         {
-            return 10f;
+            return 10.0;
         }
 
-        protected override float BuildWalkingCycleLength()
+        protected override double BuildWalkingCycleLength()
         {
-            return 5f;
+            return 5;
         }
 
-        protected override float BuildWalkingAcceleration()
+        protected override double BuildWalkingAcceleration()
         {
-            return 0.01f;
+            return 0.01;
         }
 
-        protected override float BuildMaxWalkingSpeed()
+        protected override double BuildMaxWalkingSpeed()
         {
-            return 0.30f;
+            return 0.30;
         }
 
-        protected override float BuildMaxRunningSpeed()
+        protected override double BuildMaxRunningSpeed()
         {
-            return 0.60f;
+            return 0.60;
         }
 
-        protected override float BuildStartingJumpAcceleration()
+        protected override double BuildStartingJumpAcceleration()
         {
-            return 25.0f;
+            return 25.0;
         }
 
-        protected override float BuildAttackingTime()
+        protected override double BuildAttackingTime()
         {
-            return 4f;
+            return 4;
         }
 
-        protected override float BuildWidth(Random random)
+        protected override double BuildWidth(Random random)
         {
-            return 1f;
+            return 1.0;
         }
 
-        protected override float BuildHeight(Random random)
+        protected override double BuildHeight(Random random)
         {
-            return 1f;
+            return 1.0;
         }
 
-        protected override float BuildMaxHealth()
+        protected override double BuildMaxHealth()
         {
-            return 0.5f;
+            return 0.5;
         }
 
-        protected override float BuildSafeDistanceAi()
+        protected override double BuildSafeDistanceAi()
         {
-            return 0f;
+            return 0.0;
         }
 
         protected override bool BuildIsJumpableOn()
@@ -116,24 +116,24 @@ namespace AbrahmanAdventure.sprites
             return true;
         }
 
-        protected override float BuildJumpProbability()
+        protected override double BuildJumpProbability()
         {
-            return 1.2f;
+            return 1.2;
         }
 
-        protected override float BuildHitTime()
+        protected override double BuildHitTime()
         {
-            return 16f;
+            return 16;
         }
 
-        protected override float BuildAttackStrengthCollision()
+        protected override double BuildAttackStrengthCollision()
         {
-            return 0.5f;
+            return 0.5;
         }
 
-        protected override float BuildChangeDirectionNoAiCycleLength()
+        protected override double BuildChangeDirectionNoAiCycleLength()
         {
-            return 100f;
+            return 100;
         }
 
         protected override bool BuildIsDieOnTouchGround()
@@ -220,11 +220,11 @@ namespace AbrahmanAdventure.sprites
         /// Get the sprite's current surface
         /// </summary>
         /// <returns>sprite's current surface</returns>
-        public override Surface GetCurrentSurface(out float xOffset, out float yOffset)
+        public override Surface GetCurrentSurface(out double xOffset, out double yOffset)
         {
-            xOffset = 0f;
-            yOffset = 0f;
-            int cycleDivision = WalkingCycle.GetCycleDivision(2.0f);
+            xOffset = 0;
+            yOffset = 0;
+            int cycleDivision = WalkingCycle.GetCycleDivision(2.0);
 
             if (!IsAlive)
                 return GetDeadSurface();
